Refactored data types and adjusted message length handling
- Updated `Severity`, `Facility`, `Priority`, and `ProtoVersion` types
  to `uint8` for consistency and memory optimization
- Changed `MsgLength` type from `int` to `int32` for better alignment
  with protocol specifications
- Reorganized `LogMsg` struct fields for optimal memory layout
- Adjusted related parser methods and tests to reflect type changes

+++ rfc3164/rfc3164.go
@@ -98,7 +98,7 @@ func (r *rfc3164) ParseReader(reader io.Reader) (parsesyslog.LogMsg, error) {
 			return logMessage, fmt.Errorf("failed to write bytes: %w", err)
 		}
 	}
-	logMessage.MsgLength = logMessage.Message.Len()
+	logMessage.MsgLength = int32(logMessage.Message.Len())
 
 	return logMessage, nil
 }
